requirements.txt
# AI和大模型相关
openai>=1.0.0

# Web框架和API
flask>=3.0.0
flask-socketio>=5.3.0
flask-migrate>=4.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
websockets>=12.0
python-dotenv>=1.0.0
pydantic>=2.5.0

# 认证相关
flask-jwt-extended>=4.5.0
werkzeug>=3.0.0

# 数据库
sqlalchemy>=2.0.0
# aiosqlite>=0.19.0
alembic>=1.12.0

# 安全工具集成
requests>=2.31.0
aiohttp>=3.9.0

# 工具和实用程序
pyyaml>=6.0.1

# 添加 pika 用于 RabbitMQ 通信
pika

# 添加 pymysql 用于 MySQL 连接
pymysql
cryptography